moduls dic is renamed to module
