+++ src/main/java/com/codecool/tripplanner/module/TripUser.java
@@ -1,4 +1,4 @@
-package com.codecool.tripplanner.moduls;
+package com.codecool.tripplanner.module;
 
 import javax.persistence.*;
 import java.util.ArrayList;
