+++ src/main/java/com/codecool/tripplanner/module/WalkingTour.java
@@ -1,4 +1,4 @@
-package com.codecool.tripplanner.moduls;
+package com.codecool.tripplanner.module;
 
 import com.codecool.tripplanner.enums.CityName;
 import com.codecool.tripplanner.enums.Genre;
